pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "sphinxcontrib-dv-flow"
version = "0.0.1"
dependencies = [
    'docutils',
    'requests',
    'jsonpointer',
    'pyyaml',
    'dv-flow-mgr'
]
authors = [
    {name = "Matthew Ballance", email = "matt.ballance@gmail.com"},
]
description = "Sphinx docs support for DV Flow Tasks"
license = {file = "LICENSE" }

#[project.scripts]
#dfm = "dv_flow.mgr.__main__:main"

#[tool.setuptools.package-data]
#"dv_flow.mgr" = ['share/*']